Save the logged in user's picture and name for use in comments
